split tags.rs up
